# fix #2187 by passing min size through to buffer writer when encoding