--- a/src/Common/MemoryBufferWriter.cs
+++ b/src/Common/MemoryBufferWriter.cs
@@ -24,7 +24,7 @@ namespace Microsoft.AspNetCore.Internal
         private readonly int _minimumSegmentSize;
         private int _bytesWritten;
 
-        private List<byte[]> _fullSegments;
+        private List<BufferSegment> _completedSegments;
         private byte[] _currentSegment;
         private int _position;
 
@@ -78,14 +78,14 @@ namespace Microsoft.AspNetCore.Internal
 
         public void Reset()
         {
-            if (_fullSegments != null)
+            if (_completedSegments != null)
             {
-                for (var i = 0; i < _fullSegments.Count; i++)
+                for (var i = 0; i < _completedSegments.Count; i++)
                 {
-                    ArrayPool<byte>.Shared.Return(_fullSegments[i]);
+                    _completedSegments[i].Return();
                 }
 
-                _fullSegments.Clear();
+                _completedSegments.Clear();
             }
 
             if (_currentSegment != null)
@@ -120,13 +120,13 @@ namespace Microsoft.AspNetCore.Internal
 
         public void CopyTo(IBufferWriter<byte> destination)
         {
-            if (_fullSegments != null)
+            if (_completedSegments != null)
             {
-                // Copy full segments
-                var count = _fullSegments.Count;
+                // Copy completed segments
+                var count = _completedSegments.Count;
                 for (var i = 0; i < count; i++)
                 {
-                    destination.Write(_fullSegments[i]);
+                    destination.Write(_completedSegments[i].Memory.Span);
                 }
             }
 
@@ -135,9 +135,9 @@ namespace Microsoft.AspNetCore.Internal
 
         public override Task CopyToAsync(Stream destination, int bufferSize, CancellationToken cancellationToken)
         {
-            if (_fullSegments == null)
+            if (_completedSegments == null)
             {
-                // There is only one segment so write without async
+                // There is only one segment so write without awaiting.
                 return destination.WriteAsync(_currentSegment, 0, _position);
             }
 
@@ -146,43 +146,49 @@ namespace Microsoft.AspNetCore.Internal
 
         private void EnsureCapacity(int sizeHint)
         {
-            // TODO: Use sizeHint
-            if (_currentSegment != null && _position < _currentSegment.Length)
+            // This does the Right Thing. It only subtracts _position from the current segment length if it's non-null.
+            // If _currentSegment is null, it returns 0.
+            var remainingSize = _currentSegment?.Length - _position ?? 0;
+
+            // If the sizeHint is 0, any capacity will do
+            // Otherwise, the buffer must have enough space for the entire size hint, or we need to add a segment.
+            if ((sizeHint == 0 && remainingSize > 0) || (sizeHint > 0 && remainingSize >= sizeHint))
             {
                 // We have capacity in the current segment
                 return;
             }
 
-            AddSegment();
+            AddSegment(sizeHint);
         }
 
-        private void AddSegment()
+        private void AddSegment(int sizeHint = 0)
         {
             if (_currentSegment != null)
             {
                 // We're adding a segment to the list
-                if (_fullSegments == null)
+                if (_completedSegments == null)
                 {
-                    _fullSegments = new List<byte[]>();
+                    _completedSegments = new List<BufferSegment>();
                 }
 
-                _fullSegments.Add(_currentSegment);
+                _completedSegments.Add(new BufferSegment(_currentSegment, _position));
             }
 
-            _currentSegment = ArrayPool<byte>.Shared.Rent(_minimumSegmentSize);
+            // Get a new buffer using the minimum segment size, unless the size hint is larger than a single segment.
+            _currentSegment = ArrayPool<byte>.Shared.Rent(Math.Max(_minimumSegmentSize, sizeHint));
             _position = 0;
         }
 
         private async Task CopyToSlowAsync(Stream destination)
         {
-            if (_fullSegments != null)
+            if (_completedSegments != null)
             {
                 // Copy full segments                
-                var count = _fullSegments.Count;
+                var count = _completedSegments.Count;
                 for (var i = 0; i < count; i++)
                 {
-                    var segment = _fullSegments[i];
-                    await destination.WriteAsync(segment, 0, segment.Length);
+                    var segment = _completedSegments[i];
+                    await destination.WriteAsync(segment.Buffer, 0, segment.Length);
                 }
             }
 
@@ -200,14 +206,14 @@ namespace Microsoft.AspNetCore.Internal
 
             var totalWritten = 0;
 
-            if (_fullSegments != null)
+            if (_completedSegments != null)
             {
                 // Copy full segments
-                var count = _fullSegments.Count;
+                var count = _completedSegments.Count;
                 for (var i = 0; i < count; i++)
                 {
-                    var segment = _fullSegments[i];
-                    segment.CopyTo(result, totalWritten);
+                    var segment = _completedSegments[i];
+                    segment.Memory.Span.CopyTo(result.AsSpan(totalWritten));
                     totalWritten += segment.Length;
                 }
             }
@@ -229,14 +235,14 @@ namespace Microsoft.AspNetCore.Internal
 
             var totalWritten = 0;
 
-            if (_fullSegments != null)
+            if (_completedSegments != null)
             {
                 // Copy full segments
-                var count = _fullSegments.Count;
+                var count = _completedSegments.Count;
                 for (var i = 0; i < count; i++)
                 {
-                    var segment = _fullSegments[i];
-                    segment.AsSpan().CopyTo(span.Slice(totalWritten));
+                    var segment = _completedSegments[i];
+                    segment.Memory.Span.CopyTo(span.Slice(totalWritten));
                     totalWritten += segment.Length;
                 }
             }
@@ -305,6 +311,28 @@ namespace Microsoft.AspNetCore.Internal
             if (disposing)
             {
                 Reset();
+            }
+        }
+
+        /// <summary>
+        /// Holds a byte[] from the pool and a size value. Basically a Memory but guaranteed to be backed by an ArrayPool byte[], so that we know we can return it.
+        /// </summary>
+        private struct BufferSegment
+        {
+            public byte[] Buffer { get; }
+            public int Length { get; }
+
+            public ReadOnlyMemory<byte> Memory => Buffer.AsMemory(0, Length);
+
+            public BufferSegment(byte[] buffer, int size)
+            {
+                Buffer = buffer;
+                Length = size;
+            }
+
+            public void Return()
+            {
+                ArrayPool<byte>.Shared.Return(Buffer);
             }
         }
     }