--- a/test/Microsoft.AspNetCore.SignalR.Common.Tests/Internal/Protocol/Utf8BufferTextWriterTests.cs
+++ b/test/Microsoft.AspNetCore.SignalR.Common.Tests/Internal/Protocol/Utf8BufferTextWriterTests.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Buffers;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using Microsoft.AspNetCore.Internal;
 using Microsoft.AspNetCore.SignalR.Internal;
@@ -202,20 +203,21 @@ namespace Microsoft.AspNetCore.SignalR.Common.Tests.Internal.Protocol
             textWriter.Write(chars);
             textWriter.Flush();
 
-            Assert.Equal(6, bufferWriter.Segments.Count);
+            var segments = bufferWriter.GetSegments();
+            Assert.Equal(6, segments.Count);
             Assert.Equal(1, bufferWriter.Position);
 
-            Assert.Equal((byte)'H', bufferWriter.Segments[0].Span[0]);
-            Assert.Equal((byte)'e', bufferWriter.Segments[0].Span[1]);
-            Assert.Equal((byte)'l', bufferWriter.Segments[1].Span[0]);
-            Assert.Equal((byte)'l', bufferWriter.Segments[1].Span[1]);
-            Assert.Equal((byte)'o', bufferWriter.Segments[2].Span[0]);
-            Assert.Equal((byte)' ', bufferWriter.Segments[2].Span[1]);
-            Assert.Equal((byte)'w', bufferWriter.Segments[3].Span[0]);
-            Assert.Equal((byte)'o', bufferWriter.Segments[3].Span[1]);
-            Assert.Equal((byte)'r', bufferWriter.Segments[4].Span[0]);
-            Assert.Equal((byte)'l', bufferWriter.Segments[4].Span[1]);
-            Assert.Equal((byte)'d', bufferWriter.Segments[5].Span[0]);
+            Assert.Equal((byte)'H', segments[0].Span[0]);
+            Assert.Equal((byte)'e', segments[0].Span[1]);
+            Assert.Equal((byte)'l', segments[1].Span[0]);
+            Assert.Equal((byte)'l', segments[1].Span[1]);
+            Assert.Equal((byte)'o', segments[2].Span[0]);
+            Assert.Equal((byte)' ', segments[2].Span[1]);
+            Assert.Equal((byte)'w', segments[3].Span[0]);
+            Assert.Equal((byte)'o', segments[3].Span[1]);
+            Assert.Equal((byte)'r', segments[4].Span[0]);
+            Assert.Equal((byte)'l', segments[4].Span[1]);
+            Assert.Equal((byte)'d', segments[5].Span[0]);
         }
 
         [Fact]
@@ -242,35 +244,85 @@ namespace Microsoft.AspNetCore.SignalR.Common.Tests.Internal.Protocol
             Assert.Same(textWriter1, textWriter2);
         }
 
+        [Fact]
+        private void WriteMultiByteCharactersToSmallBuffers()
+        {
+            // Test string breakdown (char => UTF-8 hex values):
+            // a => 61
+            // い => E3-81-84
+            // b => 62
+            // ろ => E3-82-8D
+            // c => 63
+            // d => 64
+            // は => E3-81-AF
+            // に => E3-81-AB
+            // e => 65
+            // ほ => E3-81-BB
+            // f => 66
+            // へ => E3-81-B8
+            // ど => E3-81-A9
+            // g => 67
+            const string testString = "aいbろcdはにeほfへどg";
+
+            // By mixing single byte and multi-byte characters, we know that there will
+            // be spaces in the active segment that cannot fit the current character. This
+            // means we'll be testing the GetMemory(minimumSize) logic.
+            var bufferWriter = new TestMemoryBufferWriter(segmentSize: 5);
+
+            var writer = new Utf8BufferTextWriter();
+            writer.SetWriter(bufferWriter);
+            writer.Write(testString);
+            writer.Flush();
+
+            // Verify the output
+            var allSegments = bufferWriter.GetSegments().Select(s => s.ToArray()).ToArray();
+            Assert.Collection(allSegments,
+                seg => Assert.Equal(new byte[] { 0x61, 0xE3, 0x81, 0x84, 0x62 }, seg),  // "aいb"
+                seg => Assert.Equal(new byte[] { 0xE3, 0x82, 0x8D, 0x63, 0x64 }, seg),  // "ろcd"
+                seg => Assert.Equal(new byte[] { 0xE3, 0x81, 0xAF }, seg),              // "は"
+                seg => Assert.Equal(new byte[] { 0xE3, 0x81, 0xAB, 0x65 }, seg),        // "にe"
+                seg => Assert.Equal(new byte[] { 0xE3, 0x81, 0xBB, 0x66 }, seg),        // "ほf"
+                seg => Assert.Equal(new byte[] { 0xE3, 0x81, 0xB8 }, seg),              // "へ"
+                seg => Assert.Equal(new byte[] { 0xE3, 0x81, 0xA9, 0x67 }, seg));       // "どg"
+
+            Assert.Equal(testString, Encoding.UTF8.GetString(bufferWriter.ToArray()));
+        }
+
         private sealed class TestMemoryBufferWriter : IBufferWriter<byte>
         {
             private readonly int _segmentSize;
 
-            internal List<Memory<byte>> Segments { get; }
+            private List<Memory<byte>> _completedSegments = new List<Memory<byte>>();
+            private int _totalLength;
+
+            public Memory<byte> CurrentSegment { get; private set; }
             internal int Position { get; private set; }
 
             public TestMemoryBufferWriter(int segmentSize = 2048)
             {
                 _segmentSize = segmentSize;
-
-                Segments = new List<Memory<byte>>();
+                CurrentSegment = Memory<byte>.Empty;
             }
-
-            public Memory<byte> CurrentSegment => Segments.Count > 0 ? Segments[Segments.Count - 1] : null;
 
             public void Advance(int count)
             {
                 Position += count;
+                _totalLength += count;
             }
 
             public Memory<byte> GetMemory(int sizeHint = 0)
             {
-                // TODO: Use sizeHint
-
-                if (Segments.Count == 0 || Position == _segmentSize)
+                // Need special handling for sizeHint == 0, because for that we want to enter the if even if there are "sizeHint" (i.e. 0) bytes left :).
+                if ((sizeHint == 0 && CurrentSegment.Length == Position) || (CurrentSegment.Length - Position < sizeHint))
                 {
-                    // TODO: Rent memory from a pool
-                    Segments.Add(new Memory<byte>(new byte[_segmentSize]));
+                    if (Position > 0)
+                    {
+                        // Complete the current segment
+                        _completedSegments.Add(CurrentSegment.Slice(0, Position));
+                    }
+
+                    // Allocate a new segment and reset the position.
+                    CurrentSegment = new Memory<byte>(new byte[_segmentSize]);
                     Position = 0;
                 }
 
@@ -284,30 +336,43 @@ namespace Microsoft.AspNetCore.SignalR.Common.Tests.Internal.Protocol
 
             public byte[] ToArray()
             {
-                if (Segments.Count == 0)
+                if (CurrentSegment.IsEmpty && _completedSegments.Count == 0)
                 {
                     return Array.Empty<byte>();
                 }
 
-                var totalLength = (Segments.Count - 1) * _segmentSize;
-                totalLength += Position;
-
-                var result = new byte[totalLength];
+                var result = new byte[_totalLength];
 
                 var totalWritten = 0;
 
-                // Copy full segments
-                for (var i = 0; i < Segments.Count - 1; i++)
+                // Copy completed segments
+                foreach (var segment in _completedSegments)
                 {
-                    Segments[i].CopyTo(result.AsMemory(totalWritten, _segmentSize));
+                    segment.CopyTo(result.AsMemory(totalWritten, segment.Length));
 
-                    totalWritten += _segmentSize;
+                    totalWritten += segment.Length;
                 }
 
-                // Copy current incomplete segment
+                // Copy current segment
                 CurrentSegment.Slice(0, Position).CopyTo(result.AsMemory(totalWritten, Position));
 
                 return result;
+            }
+
+            public IList<Memory<byte>> GetSegments()
+            {
+                var list = new List<Memory<byte>>();
+                foreach (var segment in _completedSegments)
+                {
+                    list.Add(segment);
+                }
+
+                if (CurrentSegment.Length > 0)
+                {
+                    list.Add(CurrentSegment.Slice(0, Position));
+                }
+
+                return list;
             }
         }
     }