This fixes an error that occurs when a multi-byte character is about to be written by Utf8BufferTextWriter, but there isn't enough space for it in the buffer. We use [`Encoder.Convert`](https://docs.microsoft.com/en-us/dotnet/api/system.text.encoder.convert?view=netstandard-2.0#System_Text_Encoder_Convert_System_Char__System_Int32_System_Byte__System_Int32_System_Boolean_System_Int32__System_Int32__System_Boolean__) which writes as many characters as it can fit into the provided buffer. However, if there isn't even enough space for a single character, it throws. 

Consider this example of the current behavior, where the buffers are in 4-byte segments (of course the real buffers are like 2048 bytes, but for illustration...):

1. The buffer has 4 bytes of space: `[ ] [ ] [ ] [ ]`
2. We write `aい`, which takes up 4 bytes: `[0x61] [0xE3] [0x81] [0x84]`
3. We call `GetMemory(0)` and since the buffer is full, we get a new 4 byte buffer: `[ ] [ ] [ ] [ ]`
4. We try to write `はに`. Only the first character fits, so we fill 3 bytes: `[0xE3] [0x81] [0xAF] [ ]`
5. We call `GetMemory(0)` but since the buffer is not full, we get the same buffer: `[0xE3] [0x81] [0xAF] [ ]`
6. We try to write `に` (since it wasn't written before). The buffer is only 1 byte in size, so the encoder throws:

```
System.ArgumentException: The output byte buffer is too small to contain the encoded data, encoding 'Unicode (UTF-8)' fallback 'System.Text.EncoderReplacementFallback'.
Parameter name: bytes
```

The new behavior is similar, but it specifies a `sizeHint` value of `4` (the maximum number of bytes needed to encoded a UTF-8 character to `IBufferWriter.GetMemory`, which means it will force a new segment to be allocated if there isn't enough space in the current segment. So the updated logic for the same input as above is:

1. The buffer has 4 bytes of space: `[ ] [ ] [ ] [ ]`
2. We write `aい`, which takes up 4 bytes: `[0x61] [0xE3] [0x81] [0x84]`
3. We call `GetMemory(4)`; since the buffer is full, we get a new 4 byte buffer: `[ ] [ ] [ ] [ ]`
4. We try to write `はに`. Only the first character fits, so we fill 3 bytes: `[0xE3] [0x81] [0xAF] [ ]`
5. We call `GetMemory(4)`; since the buffer doesn't have 4 bytes left, we get a new buffer: `[ ] [ ] [ ] [ ]`
6. We try to write `に` (since it wasn't written before). And it fits: `[0xE3] [0x81] [0xAB] [ ]`

I'll run the hub protocol benchmarks and post the results here.

Fixes #2187