@anurse 2018-05-02
No obvious change in perf:

```
Before:

             Method |          Input | HubProtocol |        Mean |      Error |     StdDev |        Op/s |  Gen 0 | Allocated |
------------------- |--------------- |------------ |------------:|-----------:|-----------:|------------:|-------:|----------:|
  ReadSingleMessage |   FewArguments |        Json |  1,588.4 ns |   6.770 ns |   6.001 ns |   629,557.9 | 0.0095 |     976 B |
 WriteSingleMessage |   FewArguments |        Json |  1,482.0 ns |   2.552 ns |   2.131 ns |   674,769.8 | 0.0076 |     824 B |
  ReadSingleMessage | LargeArguments |        Json | 60,528.2 ns | 197.791 ns | 165.164 ns |    16,521.2 | 0.7324 |   58872 B |
 WriteSingleMessage | LargeArguments |        Json | 36,588.6 ns |  80.589 ns |  71.440 ns |    27,330.9 | 0.3052 |   24792 B |
  ReadSingleMessage |  ManyArguments |        Json |  2,847.1 ns |   9.412 ns |   8.804 ns |   351,231.4 | 0.0153 |    1536 B |
 WriteSingleMessage |  ManyArguments |        Json |  2,781.2 ns |   5.741 ns |   5.090 ns |   359,558.5 | 0.0229 |    2080 B |
  ReadSingleMessage |    NoArguments |        Json |    859.3 ns |   3.244 ns |   2.876 ns | 1,163,745.3 | 0.0057 |     600 B |
 WriteSingleMessage |    NoArguments |        Json |    716.0 ns |   1.591 ns |   1.411 ns | 1,396,726.0 | 0.0038 |     392 B |


After:

             Method |          Input | HubProtocol |        Mean |      Error |     StdDev |        Op/s |  Gen 0 | Allocated |
------------------- |--------------- |------------ |------------:|-----------:|-----------:|------------:|-------:|----------:|
  ReadSingleMessage |   FewArguments |        Json |  1,597.0 ns |  39.848 ns |  35.324 ns |   626,192.5 | 0.0095 |     976 B |
 WriteSingleMessage |   FewArguments |        Json |  1,535.0 ns |   5.302 ns |   4.428 ns |   651,461.4 | 0.0095 |     824 B |
  ReadSingleMessage | LargeArguments |        Json | 60,830.7 ns | 271.260 ns | 253.737 ns |    16,439.1 | 0.7324 |   58872 B |
 WriteSingleMessage | LargeArguments |        Json | 36,514.7 ns |  75.660 ns |  67.070 ns |    27,386.2 | 0.3052 |   24792 B |
  ReadSingleMessage |  ManyArguments |        Json |  2,808.3 ns |   9.053 ns |   8.025 ns |   356,082.0 | 0.0153 |    1536 B |
 WriteSingleMessage |  ManyArguments |        Json |  2,780.2 ns |   7.092 ns |   6.634 ns |   359,692.2 | 0.0229 |    2080 B |
  ReadSingleMessage |    NoArguments |        Json |    828.6 ns |   1.787 ns |   1.395 ns | 1,206,807.9 | 0.0067 |     600 B |
 WriteSingleMessage |    NoArguments |        Json |    729.8 ns |   2.421 ns |   2.264 ns | 1,370,279.2 | 0.0038 |     392 B |
```