@JamesNK 2018-05-03
Doesn't perf look the same as before? But yeah, no need to go byte[] -> Memory -> Span. Skip the memory step.